==> pipeline/3_random_sample.ipynb <==
# GENERIC FIRST CELL FOR DEVELOPING A NEW METHODOLOGY

# Define an expressive name for the methodology, that will be used to create the output directory
# The complete path to the output directory will be provided in OUT_DIR
METHODOLOGY_NAME='poslog/'

# Set to number of directories below project root if the notebook is in a subdirectory of the project, so you can use relative paths
SUBDIR_LEVEL = 1
if SUBDIR_LEVEL>0:
    import sys 
    new_path = '../'*SUBDIR_LEVEL
    if new_path not in sys.path:
        sys.path.append(new_path)

import os
OUT_DIR = os.path.relpath(os.path.join(os.getcwd(), '../'*SUBDIR_LEVEL, 'out', METHODOLOGY_NAME))
if not os.path.exists(OUT_DIR):
        os.makedirs(OUT_DIR)
print(f"Current output directory: '{OUT_DIR}'")

%load_ext autoreload
%autoreload 2

# Precons

INPUT_FILE=os.path.join(OUT_DIR, '2-2_examples_tagged_upos_majority.csv')

# if file does not exist, throw an error
if not os.path.exists(INPUT_FILE):
    precon='2-2_tag_comparison_correction.ipynb'
    raise FileNotFoundError(f"File '{INPUT_FILE}' not found. Run '{precon}' first.")

OUTPUT_FILE=os.path.join(OUT_DIR,'3_random_sample_wip_upos.csv')
print(f'Output file: {OUTPUT_FILE}')

RANDOM_SAMPLE_SIZE=100
RANDOM_SAMPLE_SEED=42


import pandas as pd
tagged_examples=pd.read_csv(INPUT_FILE)

from util.pos import TagComparison

def read_lists_from_df(tagged_examples):
    log_lines=tagged_examples['Example'].to_list()
    log_lines_splitted=[eval(x) for x in tagged_examples['Tokens'].to_list()]
    tag_comparisons:list[TagComparison]=[eval(x) for x in tagged_examples['TagComparison'].to_list()]
    majorities:list[list[str]]=[eval(x) for x in tagged_examples['Majority'].to_list()]
    return log_lines, log_lines_splitted, tag_comparisons, majorities

log_lines,log_lines_splitted,tag_comparisons,majorities=read_lists_from_df(tagged_examples)


## Create Random Samples

nones_sum=0
nones_in_lines=0
token_count=0
full_tagged_indices=[]
ragged_tagged_indices=[]
for majority in majorities:
    nones=majority.count(None)
    token_count+=len(majority)
    nones_sum+=nones
    if nones>0:
        nones_in_lines+=1
        ragged_tagged_indices.append(majorities.index(majority))
    else:
        full_tagged_indices.append(majorities.index(majority))

print(f"None count: {nones_sum} of {token_count} tokens ({nones_sum/token_count:.2%})")
print(f"Lines with None: {nones_in_lines} of {len(majorities)} ({nones_in_lines/len(majorities):.2%})")
print(f"Lines with full tagging: {len(full_tagged_indices)} of {len(majorities)} ({len(full_tagged_indices)/len(majorities):.2%})")

print(full_tagged_indices)

import random
random.seed(RANDOM_SAMPLE_SEED)
print(f'Full tagged indices: {full_tagged_indices}')
random_sample_indices=random.sample(full_tagged_indices, RANDOM_SAMPLE_SIZE)
print(f'Random sample indices: {random_sample_indices}')

#####
# Due to changes in preprocessing, the length of full_tagged_indices has changed, so the random sample indices also changed.
# But we manually tagged the following indices by hand (originally they were random), so we proceed with them.

hacky_sample_indices=[915, 263, 2117, 2585, 2343, 1424, 1033, 893, 4624, 324, 299, 308, 2274, 2462, 5129, 88, 1688, 4610, 2299, 4613, 2973, 88, 1659, 4608, 3396, 2195, 1627, 2230, 3041, 1032, 386, 4098, 980, 3815, 3643, 303, 439, 4608, 308, 4078, 804, 3133, 280, 417, 417, 450, 2406, 2538, 811, 2468, 993, 4608, 3892, 981, 3960, 3772, 2170, 2848, 739, 703, 2584, 1690, 4608, 4093, 2312, 2293, 2645, 568, 2205, 326, 3333, 2498, 2855, 655, 2185, 3326, 2198, 5054, 2185, 4624, 1466, 2543, 2604, 2800, 4606, 4304, 3855, 2290, 1408, 5186, 4947, 913, 464, 1102, 791, 417, 4610, 638, 4146, 774]
random_sample_indices=hacky_sample_indices

print(f'New "random" sample indices: {random_sample_indices}')

random_sample_df=tagged_examples.iloc[random_sample_indices]
random_sample_df


## Load if exists

# # TODO
# if os.path.exists(OUTPUT_FILE):
#     os.remove(OUTPUT_FILE)
#     print(f"File '{OUTPUT_FILE}' has been removed.")
# else:
#     print(f"File '{OUTPUT_FILE}' does not exist.")

import pandas as pd
random_sample_df:pd.DataFrame

if os.path.exists(OUTPUT_FILE):
    print(f"File '{OUTPUT_FILE}' already exists. Go on editing or remove it.")
    random_sample_df=pd.read_csv(OUTPUT_FILE)
else: 
    # columns: Dataset,Line,Example,Template,ClusterId,Tokens,[tagger],TagComparison,Majority
    #tagged_examples=pd.read_csv(INPUT_FILE)
    pass

random_sample_df


# rebuild lists if df changed
log_lines,log_lines_splitted,tag_comparisons,majorities=read_lists_from_df(random_sample_df)

# "GUI"

corrected=[]
if 'ManualTagging' in random_sample_df.columns:
    corrected=[eval(x) for x in random_sample_df['ManualTagging'].to_list()]
else:
    #corrected=[m.copy() for m in majorities]
    corrected=[[None]*len(m) for m in majorities]
print(corrected)


from util.gui_and_helper import GuiLogs

gui_logs:list[GuiLogs]=[]
for i, (log_line, log_line_splitted, tag_comparison, majorities_per_line, manual_tags) in enumerate(zip(log_lines, log_lines_splitted, tag_comparisons, majorities, corrected)):
    #in random sample, we want all examples
    gui_logs.append(GuiLogs(i, log_line, log_line_splitted, tag_comparison, majorities_per_line, manual_tags))

gui_logs[:10]

from util.gui_and_helper import initialize_gui
initialize_gui(tagged_examples_df=random_sample_df,
               manually_tagged_ls=corrected,
               majorities_ls=majorities,
               gui_logs_ls=gui_logs,
               output_file_str=OUTPUT_FILE,
               find_none_only=False)

%gui tk 
from util.gui_and_helper import run_gui

run_gui()

# Save if not done yet
#random_sample_df.to_csv(OUTPUT_FILE, index=False)
print(f'Saved random samples to: {OUTPUT_FILE}')
