==> pipeline/5_poslog_crf.ipynb <==
# GENERIC FIRST CELL FOR DEVELOPING A NEW METHODOLOGY

# Define an expressive name for the methodology, that will be used to create the output directory
# The complete path to the output directory will be provided in OUT_DIR
METHODOLOGY_NAME='poslog/'

# Set to number of directories below project root if the notebook is in a subdirectory of the project, so you can use relative paths
SUBDIR_LEVEL = 1
if SUBDIR_LEVEL>0:
    import sys 
    new_path = '../'*SUBDIR_LEVEL
    if new_path not in sys.path:
        sys.path.append(new_path)

import os
OUT_DIR = os.path.relpath(os.path.join(os.getcwd(), '../'*SUBDIR_LEVEL, 'out', METHODOLOGY_NAME))
if not os.path.exists(OUT_DIR):
        os.makedirs(OUT_DIR)
print(f"Current output directory: '{OUT_DIR}'")

%load_ext autoreload
%autoreload 2

POS_LOG_MODEL='pos_log_upos_crf_10k_model2'

# Precons
INPUT_FILE=os.path.join(OUT_DIR, '4_ground_truth_wip_upos.csv')

# if file does not exist, throw an error
if not os.path.exists(INPUT_FILE):
    precon='2-2_tag_comparison_correction.ipynb'
    raise FileNotFoundError(f"File '{INPUT_FILE}' not found. Run '{precon}' first.")

import pandas as pd
example_df=pd.read_csv(INPUT_FILE)
example_df

majorities=[eval(x) for x in example_df['Majority'].to_list()]
manual_tags=[eval(x) for x in example_df['ManualTagging'].to_list()]
examples_splitted=[eval(x) for x in example_df['Tokens'].to_list()]

majorities=manual_tags

nones_sum=0
nones_in_lines=0
token_count=0
full_tagged_indices=[]
ragged_tagged_indices=[]
for majority in majorities:
    nones=majority.count(None)
    token_count+=len(majority)
    nones_sum+=nones
    if nones>0:
        nones_in_lines+=1
        ragged_tagged_indices.append(majorities.index(majority))
    else:
        full_tagged_indices.append(majorities.index(majority))

print(f"None count: {nones_sum} of {token_count} tokens ({nones_sum/token_count:.2%})")
print(f"Lines with None: {nones_in_lines} of {len(majorities)} ({nones_in_lines/len(majorities):.2%})")
print(f"Lines with full tagging: {len(full_tagged_indices)} of {len(majorities)} ({len(full_tagged_indices)/len(majorities):.2%})")

print(full_tagged_indices)

tagged_sents=[]
for word_i in full_tagged_indices:
    tokens=examples_splitted[word_i]
    tags=majorities[word_i]
    tagged_sents.append(list(zip(tokens,tags)))
print(len(tagged_sents), sum([len(x) for x in tagged_sents]))

# Train/Test Split

80/20 split

TRAIN_SPLIT=0.8

train_sents=[]
test_sents=[]

shuffeled_tagged_sents=tagged_sents.copy()
import random
random.seed(42)
random.shuffle(shuffeled_tagged_sents)

split_index = int(len(shuffeled_tagged_sents) * TRAIN_SPLIT)
train_sents = shuffeled_tagged_sents[:split_index]
test_sents = shuffeled_tagged_sents[split_index:]
print(f"Train size: {len(train_sents)} ({sum([len(x) for x in train_sents])})")
print(f"Test size: {len(test_sents)} ({sum([len(x) for x in test_sents])})")

## Evaluate

def remove_none_values_from_ys(y, y_pred, skip_x=True):
    yas=[]
    yps=[]
    for i in range(len(y)):
        ya=[]
        yp=[]
        for j in range(len(y[i])):
            if y[i][j] is not None:
                if skip_x and y[i][j]=='X':
                    continue
                ya.append(y[i][j])
                yp.append(y_pred[i][j])
        yas.append(ya)
        yps.append(yp)
    return yas, yps

from sklearn.metrics import accuracy_score
from sklearn.metrics import classification_report

def print_accuracy(y, y_pred, report=False, skip_x=True)->float:
    yas, yps = remove_none_values_from_ys(y, y_pred, skip_x=skip_x)
    y_t=[tag for tags in yas for tag in tags]
    y_p=[tag for tags in yps for tag in tags]

    acc=accuracy_score(y_t, y_p)
    print(acc)
    if report:
        print(classification_report(y_t, y_p))
    return acc



from util.pos import AbstractPosTagger

import time
def tag_and_meas_time(tagger:AbstractPosTagger, X:list[list[str]])->tuple[list[list[str]], float]:
    time_start = time.time()
    y_pred=[tagger.pos_tag(x) for x in X]
    elapsed_time=time.time()-time_start
    return y_pred, elapsed_time

def tag_and_evaluate(tagger:AbstractPosTagger, X:list[list[str]], y:list[list[str]], last_pred:tuple[list[list[str]],float]|None=None, report=False, skip_x=True)->list[list[str]]:
    print(tagger.__class__.__name__)

    if last_pred is None:
        y_pred,elapsed_time=tag_and_meas_time(tagger, X)
    else:
        y_pred,elapsed_time=last_pred

    acc=print_accuracy(y, y_pred, report, skip_x)
    lines=len(X)
    tokens=sum([len(x) for x in X])
    print(f"Lines: {lines}")
    print(f"Tokens: {tokens}")
    print(f"Elapsed time: {elapsed_time}")
    print(f"Time per 1000 lines: {elapsed_time/lines*1000}")
    print(f"Time per 1,000,000 tokens: {elapsed_time/tokens*1000000}\n")
    return (y_pred, elapsed_time, acc)


from poslog import PosLogCRF
pos_log=PosLogCRF(POS_LOG_MODEL)
pos_log.train_from_tagged_sents(train_sents)

X_test=[[token for token, tag in sent] for sent in test_sents]
y_test=[[tag for token, tag in sent] for sent in test_sents]
y_pred, elapsed_time, acc=tag_and_evaluate(pos_log, X_test, y_test, report=True, skip_x=False)

## Other taggers

from poslog import AbstractPosTagger
from util.pos import NLTKPosTagger, StanzaPosTagger, SpacyPosTagger, HanTaPosTagger, TreeTaggerPosTagger
nltk_tagger=NLTKPosTagger()
stanza_tagger=StanzaPosTagger()
spacy_tagger=SpacyPosTagger()
hanta_tagger=HanTaPosTagger()
treetagger_tagger=TreeTaggerPosTagger()

# Disabling logging for TreeTagger
# Maybe you have to restart the kernel
import logging
level_before=logging.getLogger().level
logging.getLogger().setLevel(logging.WARNING)

y_pred_treetagger=tag_and_evaluate(treetagger_tagger, X_test, y_test)

logging.getLogger().setLevel(level_before)

y_pred_nltk=tag_and_evaluate(nltk_tagger, X_test, y_test)
#y_pred_stanza=tag_and_evaluate(stanza_tagger, X_test, y_test)
y_pred_spacy=tag_and_evaluate(spacy_tagger, X_test, y_test)
y_pred_hanta=tag_and_evaluate(hanta_tagger, X_test, y_test)


y_pred_stanza=tag_and_evaluate(stanza_tagger, X_test, y_test)

## Repeat tagging for better time evaluation

X_test = X_test * 10
y_test = y_test * 10

y_pred, elapsed_time, acc=tag_and_evaluate(pos_log, X_test, y_test, report=True, skip_x=False)

# Disabling logging for TreeTagger
# Maybe you have to restart the kernel
import logging
level_before=logging.getLogger().level
logging.getLogger().setLevel(logging.WARNING)

y_pred_treetagger=tag_and_evaluate(treetagger_tagger, X_test, y_test)

logging.getLogger().setLevel(level_before)

y_pred_nltk=tag_and_evaluate(nltk_tagger, X_test, y_test)
#y_pred_stanza=tag_and_evaluate(stanza_tagger, X_test, y_test)
y_pred_spacy=tag_and_evaluate(spacy_tagger, X_test, y_test)
y_pred_hanta=tag_and_evaluate(hanta_tagger, X_test, y_test)


y_pred_stanza=tag_and_evaluate(stanza_tagger, X_test, y_test)
